requirements.txt
opencv_python==4.4.0.44
torchvision==0.7.0
torch==1.6.0
matplotlib==3.3.1
pandas==1.1.3
scipy==1.2.1
numpy==1.19.1
Pillow==8.0.0
